% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{citbi_raw}
\alias{citbi_raw}
\title{Raw ciTBI Dataset}
\format{
Data frame with 30,379 rows and 26 variables
}
\usage{
citbi_raw
}
\description{
Dataset with 30,379 pediatric head trauma patients. Missing codes: 91 = Don't know, 92 = Refused.
}
\keyword{datasets}
